+++ LibBot/Services/Interfaces/ICodeDbService.cs
@@ -6,7 +6,7 @@ namespace LibBot.Services.Interfaces;
 public interface ICodeDbService
 {
     public Task CreateItemAsync(CodeDbModel item);
-    public Task<CodeDbModel> ReadItemAsync(int chatId);
+    public Task<CodeDbModel> ReadItemAsync(long chatId);
     public Task UpdateItemAsync(CodeDbModel item);
-    public Task DeleteItemAsync(int chatId);
+    public Task DeleteItemAsync(long chatId);
 }
